MatchCompiler: Make ConstStrings static const
This allows ConstString objects to be created once, instead of
recreating them every time the object is needed. Profiling revealed that
previously, the total cost of constructing the objects costed about half
of what the comparisons did, so making them static consts does make a
difference. For a small file, this reduces IR with ~2.5%.