MatchCompiler: Refactor replaceCStrings()
By making it recursive, there's not need to add a temporary string that
is later replaced. This simplifies changes in the next commit.
Also, add a test case with multiple comparisons on the same line, which
is the reason the temporary strings were added in the first place.

There is no change to the generated code.